+++ worldengine/source/images/world_image.cpp
@@ -11,8 +11,8 @@ void WorldImage::DrawImage(boost::gil::rgb8_image_t::view_t& target)
    const BiomeArrayType&    biomes   = world_.GetBiomeData();
    const SeaDepthArrayType& seaDepth = world_.GetSeaDepthData();
 
-   const uint32_t width  = biomes.shape()[1];
-   const uint32_t height = biomes.shape()[0];
+   const uint32_t width  = static_cast<uint32_t>(biomes.shape()[1]);
+   const uint32_t height = static_cast<uint32_t>(biomes.shape()[0]);
 
    for (uint32_t y = 0; y < height; y++)
    {
@@ -24,7 +24,8 @@ void WorldImage::DrawImage(boost::gil::rgb8_image_t::view_t& target)
          }
          else
          {
-            uint8_t c    = std::clamp<int>(seaDepth[y][x] * 200 + 50, 0, 255);
+            uint8_t c = std::clamp<int32_t>(
+               static_cast<int32_t>(seaDepth[y][x] * 200 + 50), 0, 255);
             target(x, y) = boost::gil::rgb8_pixel_t(0, 0, 255 - c);
          }
       }
